Authorization, sign up, sign in, logout and login

+++ app/Http/Controllers/ChirpController.php
@@ -2,11 +2,13 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 use App\Models\Chirp;
 
 class ChirpController extends Controller
 {
+    use AuthorizesRequests;
     /**
      * Display a listing of the resource.
      */
@@ -29,52 +31,42 @@ class ChirpController extends Controller
      */
     public function store(Request $request)
     {
-        $validated = $request -> validate([
+        $validated = $request->validate([
             'message' => 'required|string|max:255',
-        ],[
-            'message.required' => 'Please write something to chirp!',
-            'message.max' => 'Chirps must 255 characters or less.',
         ]);
-        Chirp::create([
-            'message' => $validated['message'],
-        ]);
-        return redirect('/') -> with('success','Your chirp has been posted!');
-    }   
 
-    /**
-     * Display the specified resource.
-     */
-    public function show(string $id)
-    {
-        //
+        // Use the authenticated user
+        auth()->user()->chirps()->create($validated);
+
+        return redirect('/')->with('success', 'Your chirp has been posted!');
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     */
     public function edit(Chirp $chirp)
     {
+        $this->authorize('update', $chirp);
+
         return view('chirps.edit', compact('chirp'));
     }
 
-    /**
-     * Update the specified resource in storage.
-     */
     public function update(Request $request, Chirp $chirp)
     {
-        $validated = $request -> validate([
+        $this->authorize('update', $chirp);
+
+        $validated = $request->validate([
             'message' => 'required|string|max:255',
-            ]);
+        ]);
+
         $chirp->update($validated);
-        return redirect('/') -> with('success','Your chirp has been updated!');
+
+        return redirect('/')->with('success', 'Chirp updated!');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     */
     public function destroy(Chirp $chirp)
     {
+        $this->authorize('delete', $chirp);
+
         $chirp->delete();
-        return redirect('/') -> with('success', 'Chirp deleted!');
+
+        return redirect('/')->with('success', 'Chirp deleted!');
     }
 }
